Multiple Events changes:
- Uniformly rename any 'listener' to 'handler'
- Update events documentation

- Add ability to execute handlers asynchronously.
The choice of sync or async can be given either at
the time of registering the handler, or overriden
at the call to emit().
Default is synchronous execution. To globally change
it to async, set EventEmitter::ASYNC = true.

- If all handlers run synchronously, the return
from emit() is a Bool. If one or more run asynchronously,
the return value is nil (unknown), because fibers are not
waited to finish.

- on() now supports once: Bool directly; once() is just a
convenience.

- Generic events (AddListenerEvent, RemoveListenerEvent)
now receive the complete wrapped Handler(T), instead of
just the proc. This allows them to inspect the handler in
detail, including its `async`, `once`, and `at` settings,
as well as access the proc stored in `handler`.

- Handlers can be removed using off() by providing any
one of: (1) Handler() object, (2) handler proc, (3) handler
proc's hash obtained by running `obj.hash`.